pyproject.toml
[tool.poetry]
name = "narocheckerbot"
version = "0.1.0"
description = ""
authors = ["chika <fumifumibot2323@gmail.com>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.11"
"ruamel.yaml" = "^0.17.32"
"discord.py" = "^2.3.1"
feedparser = "^6.0.11"

[tool.poetry.dev-dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
